Improve file type detection
For the file names '*.nomad', set the file type to 'nomad' instead of
'hcl'. For the file names '*.tf' and '*.tfvars', set the file type to
'terraform' instead of 'hcl'. For both file types, set the syntax to
'hcl'.

This is helpful in interactions with other plugins that expect e.g.
the more specific file type 'terraform' instead of the more generic
file type 'hcl' for Terraform source files.